added the website, switched to using hardhat for local node
ive switched to using hardhat for my local node since i had issues with
anvil taking forever to process a transaction.

ive also seen there are some functionality that is not core
functionality not work and depositing does not function currently. so i
am going to test the contracts more thuroghly (as i should) and use
vitest to test the contract interface.